Update fr.po
New strings translated.

Need more informations about:
msgid "battle: deterministic events"
msgstr "combat : événements personnalisés"

N/A can be 3 things... not applicable, not available, no answer. What is the use of this string? :)
https://fr.wikipedia.org/wiki/N/a
I suggest msgstr "N/C" as for "Non communiqué"... but it's not very used... the most commonly used, near the meaning of N/A is: "R à S" or "R.A.S.", meaning "Rien à signaler"...
What difference with 2103 msgid "n/a"?
Need more informations for:
msgid "Planned Gift"

From "Project Coordination and Core Development" I am not totally sure of my words ("à" would vary with what is the next word)

For future adds, instead of: msgstr ""
Could we get: msgstr "need translation"
It would be much easier for the Ctrl+F....Because there is too much msgtr "" in the file. xD (or maybe there is another way to find what do need to be translated?)